playlists_service/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["nickAnhel <anhimovn1@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.1"
alembic = "^1.13.2"
sqlalchemy = "^2.0.31"
pydantic-settings = "^2.3.4"
asyncpg = "^0.29.0"
aiohttp = "^3.9.5"
debugpy = "^1.8.5"
grpcio = "^1.66.2"
grpcio-tools = "^1.66.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
